KAFKA-14413: Separate MirrorMaker configurations for each connector
