+++ connect/mirror/src/main/java/org/apache/kafka/connect/mirror/MirrorHeartbeatConnector.java
@@ -29,7 +29,7 @@ import java.util.Collections;
 /** Emits heartbeats to Kafka.
  */
 public class MirrorHeartbeatConnector extends SourceConnector {
-    private MirrorConnectorConfig config;
+    private MirrorHeartbeatConfig config;
     private Scheduler scheduler;
     
     public MirrorHeartbeatConnector() {
@@ -37,13 +37,13 @@ public class MirrorHeartbeatConnector extends SourceConnector {
     }
 
     // visible for testing
-    MirrorHeartbeatConnector(MirrorConnectorConfig config) {
+    MirrorHeartbeatConnector(MirrorHeartbeatConfig config) {
         this.config = config;
     }
 
     @Override
     public void start(Map<String, String> props) {
-        config = new MirrorConnectorConfig(props);
+        config = new MirrorHeartbeatConfig(props);
         scheduler = new Scheduler(MirrorHeartbeatConnector.class, config.adminTimeout());
         scheduler.execute(this::createInternalTopics, "creating internal topics");
     }
@@ -71,7 +71,7 @@ public class MirrorHeartbeatConnector extends SourceConnector {
 
     @Override
     public ConfigDef config() {
-        return MirrorConnectorConfig.CONNECTOR_CONFIG_DEF;
+        return MirrorHeartbeatConfig.CONNECTOR_CONFIG_DEF;
     }
 
     @Override
