+++ connect/mirror/src/main/java/org/apache/kafka/connect/mirror/MirrorSourceMetrics.java
@@ -17,33 +17,30 @@
 package org.apache.kafka.connect.mirror;
 
 import org.apache.kafka.common.MetricNameTemplate;
+import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.metrics.Metrics;
 import org.apache.kafka.common.metrics.MetricsReporter;
 import org.apache.kafka.common.metrics.Sensor;
-import org.apache.kafka.common.metrics.stats.Value;
-import org.apache.kafka.common.metrics.stats.Min;
-import org.apache.kafka.common.metrics.stats.Max;
 import org.apache.kafka.common.metrics.stats.Avg;
+import org.apache.kafka.common.metrics.stats.Max;
 import org.apache.kafka.common.metrics.stats.Meter;
-import org.apache.kafka.common.TopicPartition;
+import org.apache.kafka.common.metrics.stats.Min;
+import org.apache.kafka.common.metrics.stats.Value;
 
 import java.util.Arrays;
-import java.util.Set;
 import java.util.HashSet;
-import java.util.Map;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 /** Metrics for replicated topic-partitions */
-class MirrorMetrics implements AutoCloseable {
+class MirrorSourceMetrics implements AutoCloseable {
 
     private static final String SOURCE_CONNECTOR_GROUP = MirrorSourceConnector.class.getSimpleName();
-    private static final String CHECKPOINT_CONNECTOR_GROUP = MirrorCheckpointConnector.class.getSimpleName();
 
     private static final Set<String> PARTITION_TAGS = new HashSet<>(Arrays.asList("target", "topic", "partition"));
-    private static final Set<String> GROUP_TAGS = new HashSet<>(Arrays.asList("source", "target", "group", "topic", "partition"));
-    
+
     private static final MetricNameTemplate RECORD_COUNT = new MetricNameTemplate(
             "record-count", SOURCE_CONNECTOR_GROUP,
             "Number of source records replicated to the target cluster.", PARTITION_TAGS);
@@ -81,27 +78,13 @@ class MirrorMetrics implements AutoCloseable {
             "replication-latency-ms-avg", SOURCE_CONNECTOR_GROUP,
             "Average time it takes records to replicate from source to target cluster.", PARTITION_TAGS);
 
-    private static final MetricNameTemplate CHECKPOINT_LATENCY = new MetricNameTemplate(
-            "checkpoint-latency-ms", CHECKPOINT_CONNECTOR_GROUP,
-            "Time it takes consumer group offsets to replicate from source to target cluster.", GROUP_TAGS);
-    private static final MetricNameTemplate CHECKPOINT_LATENCY_MAX = new MetricNameTemplate(
-            "checkpoint-latency-ms-max", CHECKPOINT_CONNECTOR_GROUP,
-            "Max time it takes consumer group offsets to replicate from source to target cluster.", GROUP_TAGS);
-    private static final MetricNameTemplate CHECKPOINT_LATENCY_MIN = new MetricNameTemplate(
-            "checkpoint-latency-ms-min", CHECKPOINT_CONNECTOR_GROUP,
-            "Min time it takes consumer group offsets to replicate from source to target cluster.", GROUP_TAGS);
-    private static final MetricNameTemplate CHECKPOINT_LATENCY_AVG = new MetricNameTemplate(
-            "checkpoint-latency-ms-avg", CHECKPOINT_CONNECTOR_GROUP,
-            "Average time it takes consumer group offsets to replicate from source to target cluster.", GROUP_TAGS);
 
-
-    private final Metrics metrics; 
-    private final Map<TopicPartition, PartitionMetrics> partitionMetrics; 
-    private final Map<String, GroupMetrics> groupMetrics = new HashMap<>();
+    private final Metrics metrics;
+    private final Map<TopicPartition, PartitionMetrics> partitionMetrics;
     private final String source;
     private final String target;
 
-    MirrorMetrics(MirrorTaskConfig taskConfig) {
+    MirrorSourceMetrics(MirrorSourceTaskConfig taskConfig) {
         this.target = taskConfig.targetClusterAlias();
         this.source = taskConfig.sourceClusterAlias();
         this.metrics = new Metrics();
@@ -140,15 +123,6 @@ class MirrorMetrics implements AutoCloseable {
         partitionMetrics.get(topicPartition).byteSensor.record((double) bytes);
     }
 
-    void checkpointLatency(TopicPartition topicPartition, String group, long millis) {
-        group(topicPartition, group).checkpointLatencySensor.record((double) millis);
-    }
-
-    GroupMetrics group(TopicPartition topicPartition, String group) {
-        return groupMetrics.computeIfAbsent(String.join("-", topicPartition.toString(), group),
-            x -> new GroupMetrics(topicPartition, group));
-    }
-
     void addReporter(MetricsReporter reporter) {
         metrics.addReporter(reporter);
     }
@@ -184,25 +158,6 @@ class MirrorMetrics implements AutoCloseable {
             replicationLatencySensor.add(metrics.metricInstance(REPLICATION_LATENCY_MAX, tags), new Max());
             replicationLatencySensor.add(metrics.metricInstance(REPLICATION_LATENCY_MIN, tags), new Min());
             replicationLatencySensor.add(metrics.metricInstance(REPLICATION_LATENCY_AVG, tags), new Avg());
-        }
-    }
-
-    private class GroupMetrics {
-        private final Sensor checkpointLatencySensor;
-
-        GroupMetrics(TopicPartition topicPartition, String group) {
-            Map<String, String> tags = new LinkedHashMap<>();
-            tags.put("source", source); 
-            tags.put("target", target); 
-            tags.put("group", group);
-            tags.put("topic", topicPartition.topic());
-            tags.put("partition", Integer.toString(topicPartition.partition()));
- 
-            checkpointLatencySensor = metrics.sensor("checkpoint-latency");
-            checkpointLatencySensor.add(metrics.metricInstance(CHECKPOINT_LATENCY, tags), new Value());
-            checkpointLatencySensor.add(metrics.metricInstance(CHECKPOINT_LATENCY_MAX, tags), new Max());
-            checkpointLatencySensor.add(metrics.metricInstance(CHECKPOINT_LATENCY_MIN, tags), new Min());
-            checkpointLatencySensor.add(metrics.metricInstance(CHECKPOINT_LATENCY_AVG, tags), new Avg());
         }
     }
 }
