+++ connect/mirror/src/test/java/org/apache/kafka/connect/mirror/MirrorMakerConfigTest.java
@@ -143,23 +143,26 @@ public class MirrorMakerConfigTest {
         SourceAndTarget sourceAndTarget = new SourceAndTarget("source", "target");
         Map<String, String> connectorProps = mirrorConfig.connectorBaseConfig(sourceAndTarget,
             MirrorSourceConnector.class);
-        MirrorConnectorConfig connectorConfig = new MirrorConnectorConfig(connectorProps);
-        assertEquals(100, (int) connectorConfig.getInt("tasks.max"),
+        MirrorSourceConfig sourceConfig = new MirrorSourceConfig(connectorProps);
+        assertEquals(100, (int) sourceConfig.getInt("tasks.max"),
             "Connector properties like tasks.max should be passed through to underlying Connectors.");
-        assertEquals(Collections.singletonList("topic-1"), connectorConfig.getList("topics"),
+        assertEquals(Collections.singletonList("topic-1"), sourceConfig.getList("topics"),
             "Topics include should be passed through to underlying Connectors.");
-        assertEquals(Collections.singletonList("group-2"), connectorConfig.getList("groups"),
+        assertEquals(Collections.singletonList("property-3"), sourceConfig.getList("config.properties.exclude"),
+                "Config properties exclude should be passed through to underlying Connectors.");
+        assertEquals(Collections.singletonList("FakeMetricsReporter"), sourceConfig.getList("metric.reporters"),
+                "Metrics reporters should be passed through to underlying Connectors.");
+        assertEquals("DefaultTopicFilter", sourceConfig.getClass("topic.filter.class").getSimpleName(),
+                "Filters should be passed through to underlying Connectors.");
+        assertEquals("__", sourceConfig.getString("replication.policy.separator"),
+                "replication policy separator should be passed through to underlying Connectors.");
+        assertFalse(sourceConfig.originals().containsKey("xxx"),
+                "Unknown properties should not be passed through to Connectors.");
+
+        MirrorCheckpointConfig checkpointConfig = new MirrorCheckpointConfig(connectorProps);
+        assertEquals(Collections.singletonList("group-2"), checkpointConfig.getList("groups"),
             "Groups include should be passed through to underlying Connectors.");
-        assertEquals(Collections.singletonList("property-3"), connectorConfig.getList("config.properties.exclude"),
-            "Config properties exclude should be passed through to underlying Connectors.");
-        assertEquals(Collections.singletonList("FakeMetricsReporter"), connectorConfig.getList("metric.reporters"),
-            "Metrics reporters should be passed through to underlying Connectors.");
-        assertEquals("DefaultTopicFilter", connectorConfig.getClass("topic.filter.class").getSimpleName(),
-            "Filters should be passed through to underlying Connectors.");
-        assertEquals("__", connectorConfig.getString("replication.policy.separator"),
-            "replication policy separator should be passed through to underlying Connectors.");
-        assertFalse(connectorConfig.originals().containsKey("xxx"),
-            "Unknown properties should not be passed through to Connectors.");
+
     }
 
     @Test
@@ -173,18 +176,19 @@ public class MirrorMakerConfigTest {
         SourceAndTarget sourceAndTarget = new SourceAndTarget("source", "target");
         Map<String, String> connectorProps = mirrorConfig.connectorBaseConfig(sourceAndTarget,
                                                                               MirrorSourceConnector.class);
-        MirrorConnectorConfig connectorConfig = new MirrorConnectorConfig(connectorProps);
+        MirrorSourceConfig sourceConfig = new MirrorSourceConfig(connectorProps);
         DefaultTopicFilter.TopicFilterConfig filterConfig =
             new DefaultTopicFilter.TopicFilterConfig(connectorProps);
 
         assertEquals(Collections.singletonList("topic3"), filterConfig.getList("topics.exclude"),
             "Topics exclude should be backwards compatible.");
 
-        assertEquals(Collections.singletonList("group-7"), connectorConfig.getList("groups.exclude"),
-            "Groups exclude should be backwards compatible.");
-
-        assertEquals(Collections.singletonList("property-3"), connectorConfig.getList("config.properties.exclude"),
+        assertEquals(Collections.singletonList("property-3"), sourceConfig.getList("config.properties.exclude"),
             "Config properties exclude should be backwards compatible.");
+
+        MirrorCheckpointConfig checkpointConfig = new MirrorCheckpointConfig(connectorProps);
+        assertEquals(Collections.singletonList("group-7"), checkpointConfig.getList("groups.exclude"),
+            "Groups exclude should be backwards compatible.");
 
     }
 
@@ -198,7 +202,7 @@ public class MirrorMakerConfigTest {
         SourceAndTarget sourceAndTarget = new SourceAndTarget("source", "target");
         Map<String, String> connectorProps = mirrorConfig.connectorBaseConfig(sourceAndTarget,
                                                                               MirrorSourceConnector.class);
-        MirrorConnectorConfig connectorConfig = new MirrorConnectorConfig(connectorProps);
+        MirrorCheckpointConfig connectorConfig = new MirrorCheckpointConfig(connectorProps);
         DefaultTopicFilter.TopicFilterConfig filterConfig =
             new DefaultTopicFilter.TopicFilterConfig(connectorProps);
 
@@ -355,6 +359,17 @@ public class MirrorMakerConfigTest {
         ConfigException ce = assertThrows(ConfigException.class,
                 () -> new MirrorClientConfig(makeProps(CommonClientConfigs.SECURITY_PROTOCOL_CONFIG, "abc")));
         assertTrue(ce.getMessage().contains(CommonClientConfigs.SECURITY_PROTOCOL_CONFIG));
+    }
+
+    @Test
+    public void testAllConfigNames() {
+        MirrorMakerConfig mirrorConfig = new MirrorMakerConfig(makeProps(
+                "clusters", "a, b"));
+        Set<String> allNames = mirrorConfig.allConfigNames();
+
+        assertTrue(allNames.contains("topics"));
+        assertTrue(allNames.contains("groups"));
+        assertTrue(allNames.contains("emit.heartbeats.enabled"));
     }
 
     public static class FakeConfigProvider implements ConfigProvider {
